refactor: import appearance names from the package
Until now a caller had to know which file inside the services and
view models directories holds a class. That puts the file layout
into every import, so renaming or splitting a module turns into a
wide change.

Going through the package instead matches how the models directory
and the shared services package already work, so all of them read
the same way.
