+++ mpvqc/appearance/services/__init__.py
@@ -1,3 +1,8 @@
 # SPDX-FileCopyrightText: mpvQC developers
 #
 # SPDX-License-Identifier: GPL-3.0-or-later
+
+from .color_scheme import ColorSchemeService as ColorSchemeService
+from .color_scheme import QtStyleHints as QtStyleHints
+from .palette_catalog import PaletteCatalogService as PaletteCatalogService
+from .palette_catalog import PaletteFamily as PaletteFamily
