+++ mpvqc/appearance/viewmodels/__init__.py
@@ -2,5 +2,11 @@
 #
 # SPDX-License-Identifier: GPL-3.0-or-later
 
+from .dialog import AppearanceDialogInputs as AppearanceDialogInputs
+from .dialog import AppearanceDialogProps as AppearanceDialogProps
 from .dialog import MpvqcAppearanceDialogViewModel as MpvqcAppearanceDialogViewModel
+from .dialog import derive_appearance_dialog_props as derive_appearance_dialog_props
 from .palette import MpvqcPaletteViewModel as MpvqcPaletteViewModel
+from .palette import PaletteInputs as PaletteInputs
+from .palette import PaletteProps as PaletteProps
+from .palette import derive_palette_props as derive_palette_props
