+++ test/appearance/viewmodels/test_palette.py
@@ -19,9 +19,8 @@ from mpvqc.appearance.domain import (
     Palette,
     Unknown,
 )
-from mpvqc.appearance.services.color_scheme import ColorSchemeService
-from mpvqc.appearance.services.palette_catalog import PaletteCatalogService
-from mpvqc.appearance.viewmodels.palette import (
+from mpvqc.appearance.services import ColorSchemeService, PaletteCatalogService
+from mpvqc.appearance.viewmodels import (
     MpvqcPaletteViewModel,
     PaletteInputs,
     PaletteProps,
